Further refactor, update to latest commit on ESLint PR
