+++ lib/babylon-to-espree/attachComments.js
