+++ lib/babylon-to-espree/convertTemplateType.js
