+++ lib/babylon-to-espree/toToken.js
