+++ lib/make-enhanced-referencer.js
