+++ lib/shared.js
